Fixed some errors that would be printed while running tests that bind to socket paths
Motivation:

Binding HTTPBin to a socket path generates many warnings in the console that are largely inconsequential, but cloud the results. This removes one of them 😅 (I'm not sure how to address the others)

Modifications:

TCP_NODELAY is now a default option in NIO, so we can leave out the setting completely since it isn't supported when binding to socket paths.

Result:

One fewer warning in the console 🙃