Fixed an issue where supplying an empty host to the Client TLS handler would result in an SNI error
Motivation:

Testing on Linux seemed to reveal an `NIOSSLExtraError.invalidSNIHostname` error that was thrown when using `https+unix` URLs, which don't have their hostname filled in. The same NIO failure does not occur on macOS, as it seems that Network.framework does not mind an empty hostname.

Modifications:

Added an extra check to determine if the host is empty before passing it to a `NIOSSLClientHandler`.

Result:

testHTTPSPlusUNIX() no longer fails on Linux.