Allow redirect configuration per request
Motivation:

Redirect configuration is set globally when configuring the `HTTPClient`. However, it would also be useful to configure redirects on a per request basis. This would mean that a single `HTTPClient` could be configured and shared widely throughout an application, while retaining the flexibility to tailor specific requests with custom redirect handling should the need arise.

Modifications:

Adds new methods providing a `redirectConfiguration` argument to support per request redirection.

When the request is executed the `prevailingConfiguration` is determined using the `redirectConfiguration` argument, if set. Otherwise the global configuration is used.

Fixes code formatting.

Result:

Allows redirect handling to be configured on a per request basis, closes #196