Fixes the DelayOnHeadDelegate test utility (#374)
Currently the DelayOnHeadDelegate test utility depends on correct timings. Right now, if a request is slower than 50ms in  `testErrorAfterCloseWhileBackpressureExerted` the test will fail, since the backpressure promise is failed, before a head was received. This pr fixes this, by giving the user a callback, when the head was received.