Remove tests relying on OS-dependent behaviour
Motivation

URL parsing has changed in macOS Sonoma and associated releases. Our
tests were reliant on the old behaviour. The behaviour is controlled
by the OS on which the program was linked, which makes it impossible
for us to programmatically work out which result we should see.

The affected tests are not actually useful, we don't really care how
the URLs parse, so we can safely just remove them.

Modifications

Remove the affected tests.

Result

Tests pass!