Update async usages in README and examples to shutdown properly
Motivation:

In several examples of async/await support, the call to shutdown() on
the HTTPClient are made with await but do not call get() on the
EventLoopFuture.

Modifications:

Add the .get() call in four locations that need it.

Result:

After fixing this, the client should properly get closed down and
there are no longer warnings emitted by the compiler.