Add an option to enable Multipath TCP on clients
Multipath TCP (MPTCP) is a TCP extension allowing to enhance
the reliability of the network by using multiple interfaces.
This extension provides a seamless handover between interfaces
in case of deterioration of the connection on the original one.
In the context of iOS and Mac OS X, it could be really interesting
to leverage the capabilities of MPTCP as they could benefit from
their multiple interfaces (ethernet + Wi-fi for Mac OS X, Wi-fi +
cellular for iOS).

This contribution introduces patches to HTTPClient.Configuration and
establishment of the Bootstraps. A supplementary field "enableMultipath"
was added to the configuration, allowing to request the use of MPTCP.
This flag is then used when creating the channels to configure the client.

Note that in the future, it might also be potentially interesting to offer
more precise configuration options for MPTCP on MacOS, as the Network
framework allows also to select a type of service, instead of just offering
the option to create MPTCP connections. Currently, when enabling MPTCP, only
the Handover mode is used.