Avoid precondition failure in write timeout
Motivation:

In some cases we can crash because of a precondition failure when the
write timeout fires and we aren't in the running state. This can happen
for example if the connection is closed whilst the write timer is
active.

Modifications:

Remove the precondition and instead take no action if the timeout fires
outside of the running state. Instead we take a new `Action`,
`.noAction` when the timer fires.

Result:

Fewer crashes.