Add missing import of Network module
In https://github.com/swift-server/async-http-client/pull/794, the new Swift
compiler feature `MemberImportVisibility` was adopted. However, CI for this
project does not test building the package with the nightly compiler on macOS,
so it was missed that the project no longer builds on macOS with the nightly
toolchain due to a missing import.