Clean up Task error handling.
Motivation

We have some Task error handling functions that are generic for
no apparent reason. They're also typically called from contexts
where they also report the error to the delegate, but one of the
call sites doesn't do that. So add a test for that as well.

Modifications

- Rewrite Task.fail(with:delegate:) to be non-generic.
- Add a call to the delegate error handler on the path that is
    missing it.
- Add a test for that call

Results

Cleaner, easier to follow code