Don't hold a lock over a continuation in test helpers (#872)
Motivation:

The various 'withMumbleContinuation' APIs are supposed to be invoked
synchronously with the caller. This assumption allows a lock to be
acquired before the call and released from the body of the
'withMumbleContinuation' after e.g. storing the continuation. However
this isn't the case and the job may be re-enqueued on the executor
meaning that this is pattern is vulnerable to deadlocks.

Modifications:

- Rework the test helpers to avoid holding a lock when a continuation is
created.
- Switch to using NIOLockedValue box

Result:

Lower chance of deadlock